+++ src/deck-components/Button.tsx
@@ -1,30 +1,21 @@
-import { CSSProperties, FC } from 'react';
+import { CSSProperties, FC, RefAttributes } from 'react';
 
 import { CommonUIModule } from '../webpack';
 
-export interface ButtonProps {
-  style: CSSProperties;
+export interface DialogButtonProps extends RefAttributes<HTMLDivElement> {
+  label?: string;
+  style?: CSSProperties;
   className?: string;
   noFocusRing?: boolean;
-  disabled?: boolean;
+  description?: string;
+  layout?: 'below';
   onClick?(e: MouseEvent): void;
-  onPointerDown?(e: PointerEvent): void;
-  onPointerUp?(e: PointerEvent): void;
-  onPointerCancel?(e: PointerEvent): void;
-  onMouseDown?(e: PointerEvent): void;
-  onMouseUp?(e: MouseEvent): void;
-  onTouchStart?(e: TouchEvent): void;
-  onTouchEnd?(e: TouchEvent): void;
-  onTouchCancel?(e: TouchEvent): void;
-  onSubmit?(e: SubmitEvent): void;
+  disabled?: boolean;
+  bottomSeparator?: boolean;
 }
 
 export const DialogButton = Object.values(CommonUIModule).find(
   (mod: any) =>
     mod?.render?.toString()?.includes('Object.assign({type:"button"') &&
     mod?.render?.toString()?.includes('DialogButton'),
-) as any;
-
-// Button isn't exported, so call DialogButton to grab it
-
-export const Button = DialogButton!.render({}).type as FC<ButtonProps>; // its actually a forwarded ref but that doesn't really matter in usage
+) as FC<DialogButtonProps>;